Latest

+++ home/dot_local/bin/post-installx/executable_post-cloudflared.sh
@@ -1,6 +1,8 @@
 #!/usr/bin/env bash
 # @file Cloudflared Configuration
 # @brief Applies cloudflared configuration, connects to Argo tunnel with managed configuration, and enables it on system start
+# @description
+#     1. Skips the deletion of a tunnel when it is currently in use
 
 if command -v cloudflared > /dev/null; then
   # Show warning message about ~/.cloudflared already existing
@@ -17,19 +19,34 @@ if command -v cloudflared > /dev/null; then
   ### Remove previous tunnels connected to host
   while read TUNNEL_ID; do
     logg info "Deleteing CloudFlared tunnel ID $TUNNEL_ID"
-    sudo cloudflared tunnel delete "$TUNNEL_ID"
-    sudo rm -f "/usr/local/etc/cloudflared/${TUNNEL_ID}.json"
+    unset TUNNEL_EXIT_CODE
+    sudo cloudflared tunnel delete "$TUNNEL_ID" || TUNNEL_EXIT_CODE=$?
+    if [ -z "$TUNNEL_EXIT_CODE" ]; then
+      logg info "Removing credentials for $TUNNEL_ID which is not in use"
+      sudo rm -f "/usr/local/etc/cloudflared/${TUNNEL_ID}.json"
+    else
+      logg success "Skipping deletion of $TUNNEL_ID credentials since it is in use"
+    fi
   done< <(sudo cloudflared tunnel list | grep "host-$HOSTNAME" | sed 's/ .*//')
 
   ### Register tunnel (if not already registered)
   logg info "Creating CloudFlared tunnel named host-$HOSTNAME"
   sudo cloudflared tunnel create "host-$HOSTNAME"
 
+  ### Acquire TUNNEL_ID and symlink credentials.json
   TUNNEL_ID="$(sudo cloudflared tunnel list | grep "host-$HOSTNAME" | sed 's/ .*//')"
   logg info "Tunnel ID: $TUNNEL_ID"
   logg info "Symlinking /usr/local/etc/cloudflared/$TUNNEL_ID.json to /usr/local/etc/cloudflared/credentials.json"
   sudo rm -f /usr/local/etc/cloudflared/credentials.json
   sudo ln -s /usr/local/etc/cloudflared/$TUNNEL_ID.json /usr/local/etc/cloudflared/credentials.json
+
+  ### Configure DNS
+  # Must be deleted manually if no longer used
+  logg info 'Setting up DNS records for CloudFlare Argo tunnels'
+  while read DOMAIN; do
+    logg info "Setting up $DOMAIN for access through cloudflared"
+    sudo cloudflared tunnel route dns "$TUNNEL_ID" "$DOMAIN" && logg success "Successfully routed $DOMAIN to this machine's cloudflared Argo tunnel"
+  done< <(yq '.ingress[].hostname' config.yml)
 
   ### Set up service
   if [ -d /Applications ] && [ -d /System ]; then
